Fix for rake not working in production

+++ lib/app_env/db/rake_tasks.rb
@@ -3,22 +3,24 @@
 require "rom/sql/rake_task"
 require 'sequel'
 
-require_relative 'config'
-app_db_config = AppEnv::DB::Config.new
+unless AppEnv.production?
+  require_relative 'config'
+  app_db_config = AppEnv::DB::Config.new
 
-namespace :db do
-  desc 'Create database'
-  task :create do
-    Sequel.connect(app_db_config.to_h.merge(database: :postgres)) do |db|
-      db.execute "DROP DATABASE IF EXISTS #{app_db_config.database}"
-      db.execute "CREATE DATABASE #{app_db_config.database}"
+  namespace :db do
+    desc 'Create database'
+    task :create do
+      Sequel.connect(app_db_config.to_h.merge(database: :postgres)) do |db|
+        db.execute "DROP DATABASE IF EXISTS #{app_db_config.database}"
+        db.execute "CREATE DATABASE #{app_db_config.database}"
+      end
     end
-  end
 
-  desc 'Drop database'
-  task :drop do
-    Sequel.connect(app_db_config.to_h.merge(database: :postgres)) do |db|
-      db.execute "DROP DATABASE IF EXISTS #{app_db_config.database}"
+    desc 'Drop database'
+    task :drop do
+      Sequel.connect(app_db_config.to_h.merge(database: :postgres)) do |db|
+        db.execute "DROP DATABASE IF EXISTS #{app_db_config.database}"
+      end
     end
   end
 end
